v 2.12 fix: control de stock e inventario, cinco defectos de la revision final
- Cancelar/reactivar un pedido de forma repetida inflaba el stock sin limite:
  ahora una transicion desde cancelado/rechazado hacia un estado no terminal
  se rechaza con 409 (InvalidStatusTransitionError), preservando la regla ya
  documentada de que un pedido cancelado no vuelve atras.
- Un pedido creado directamente en estado cancelado/rechazado descontaba
  stock que despues nunca se devolvia (deleteOrder asume que ya se restauro).
  createOrder ahora salta la validacion de faltantes y el descuento cuando el
  estado inicial ya es terminal, igual que ya hacia deleteOrder.
- unitCost coercionaba null/""/false a 0 en vez de rechazarlos, escondiendo
  la falta de costo cargado. Se saco el z.coerce.
- Un id de ingrediente repetido en POST /dish generaba un 500 por clave
  duplicada; ahora se dedupea en el schema igual que ya hace updateDish,
  quedandose con la ultima cantidad.
- Una cantidad fraccionaria en un item de pedido corrompia el stock entero;
  se agrego .int() igual que ya tiene DishIngredientSchema.

+++ src/order/order.schema.ts
@@ -7,7 +7,7 @@ import { OrderStatus } from '../shared/enum/order.statusEnum.js';
 
 export const OrderItemSchema = z.object({
   dish: DishIdSchema.transform((obj) => obj.id),
-  quantity: z.coerce.number().pipe(z.number().min(1)),
+  quantity: z.coerce.number().pipe(z.number().int().min(1)),
 });
 
 export const OrderItemIdSchema = z.object({
